// src/stores/bank-garansi/models/IBankGaransiState.ts
import BankGaransiModel from './BankGaransiModel';
import BankGaransiEditViewAdminModel from './BankGaransiEditViewAdminModel';
import BankGaransiEditViewRequesterModel from './BankGaransiEditViewRequesterModel';
import FunnelSATableModel from './FunnelSATableModel';
import FunnelSALinkToModel from './FunnelSALinkToModel';
import BankGaransiAdminEnvelope from './BankGaransiAdminEnvelope';
import BankGaransiEnvelope from './BankGaransiEnvelope';
import BankRecommended from './BankRecommended';
import DropdownFunnelSAModel from './DropdownFunnelSAModel';
import AttachmentEnvelope from './AttachmentEnvelope';
import MasterInsuranceUdcModel from './MasterInsuranceUdcModel';
import FilterSearchModel from './FilterSearchModel';
import MasterInsuranceEnvelope from './MasterInsuranceEnvelope';
import CompetitorProductModel from 'stores/competitor-product/models/CompetitorProductModel';
import FunnelSARowModel from './FunnelSARowModel';
import FunnelPORowModel from './FunnelPORowModel';
import BankGaransiDashboardEnvelope from './BankGaransiDashboardEnvelope';
import BankGaransiActivityModel from './BankGaransiActivityModel';
import FunnelPOTableModel from './FunnelPOTableModel';
import ExtendAttachmentModel from './ExtendAttachmentModel';
import AttachmentVersionEnvelope from './AttachmentVersionEnvelope';
import MaxAmountModel from './MaxAmountModel';
import CheckExpireModel from './CheckExpireModel';

export default interface IBankGaransiState {
  readonly data: BankGaransiModel[];
  readonly dataAttachment: AttachmentEnvelope;
  readonly dataDivision: CompetitorProductModel[];
  readonly dataInsurance: MasterInsuranceEnvelope;
  readonly dataPrint: CompetitorProductModel[];
  readonly dropdownFunnelSA: DropdownFunnelSAModel[];
  readonly customerBG: DropdownFunnelSAModel[];
  readonly creatorBG: DropdownFunnelSAModel[];
  readonly firstData: BankGaransiModel;
  readonly firstDataInsurance: MasterInsuranceUdcModel;
  readonly firstDataAdmin: BankGaransiEditViewAdminModel;
  readonly requesterData: BankGaransiEditViewRequesterModel;
  readonly dataFunnelSA: FunnelSATableModel;
  readonly dataFunnelPO: FunnelPOTableModel;
  readonly dataLinkTo: FunnelSALinkToModel;
  readonly extendAttachment: ExtendAttachmentModel;
  readonly listData: BankGaransiEnvelope;
  readonly listExpiredData: BankGaransiEnvelope;
  readonly listDataAdmin: BankGaransiAdminEnvelope;
  readonly listExpiredDataAdmin: BankGaransiDashboardEnvelope;
  readonly listSearch: BankGaransiDashboardEnvelope;
  readonly bankRecomended: BankRecommended;
  readonly bankEstimated: BankRecommended;
  readonly error: boolean;
  readonly refreshPage: boolean;
  readonly refreshAdminPage: boolean;
  readonly refreshRequesterPage: boolean;
  readonly dataSearch: FilterSearchModel;
  readonly dataCompany: MasterInsuranceUdcModel[];
  readonly dataCompanyApplicant: MasterInsuranceUdcModel[];
  readonly dataBankCG: DropdownFunnelSAModel[];
  readonly funnelSAObject: FunnelSARowModel;
  readonly funnelPOObject: FunnelPORowModel;
  readonly resultActions: any;
  readonly mandatoryAttach: any;
  readonly maxAmount: MaxAmountModel;
  readonly expired: CheckExpireModel;
  readonly listHistory: BankGaransiActivityModel[];
  readonly attachmentVersion: AttachmentVersionEnvelope;
  readonly activePage: number;
}